Move the implementation of ExceptionDetails to a third partial class
