+++ src/Core/Managed/Shared/Extensibility/Implementation/External/ExceptionDetails_types.cs
@@ -26,7 +26,7 @@
 
 namespace Microsoft.ApplicationInsights.Extensibility.Implementation.External
 {
-    using System;
+    using System.Collections.Concurrent;
     using System.Collections.Generic;
 
 
@@ -76,31 +76,6 @@ namespace Microsoft.ApplicationInsights.Extensibility.Implementation.External
             hasFullStack = true;
             stack = "";
             parsedStack = new List<StackFrame>();
-        }
-
-        /// <summary>
-        /// Creates a new instance of ExceptionDetails from a System.Exception and a parent ExceptionDetails.
-        /// </summary>
-        internal static ExceptionDetails CreateWithoutStackInfo(Exception exception, ExceptionDetails parentExceptionDetails)
-        {
-            if (exception == null)
-            {
-                throw new ArgumentNullException("exception");
-            }
-
-            var exceptionDetails = new External.ExceptionDetails()
-            {
-                id = exception.GetHashCode(),
-                typeName = exception.GetType().FullName,
-                message = exception.Message
-            };
-
-            if (parentExceptionDetails != null)
-            {
-                exceptionDetails.outerId = parentExceptionDetails.id;
-            }
-
-            return exceptionDetails;
         }
     }
 } // AI
